Revert "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
This reverts commit da0a3d19e2f4b187d801f401e884915ee3a052b0.

+++ code/modules/explorer_drone/controle_console.dm
@@ -63,9 +63,7 @@
 		.["configurable"] = controlled_drone.drone_status == EXODRONE_IDLE
 		.["cargo"] = controlled_drone.get_cargo_data()
 		.["drone_travel_coefficent"] = controlled_drone.get_travel_coeff()
-		var/travel_error = controlled_drone.travel_error()
-		.["can_travel"] = !travel_error
-		.["travel_error"] = travel_error ? travel_error : ""
+		.["can_travel"] = controlled_drone.can_travel()
 		switch(controlled_drone.drone_status) //could move this down to drone.ui_data()
 			if(EXODRONE_IDLE)
 				.["sites"] = build_exploration_site_ui_data()
@@ -113,7 +111,8 @@
 			signal_lost = FALSE
 			return TRUE
 		if("self_destruct")
-			qdel(controlled_drone) //var will be nulled in signal response
+			if(controlled_drone)
+				qdel(controlled_drone) //var will be nulled in signal response
 			return TRUE
 		if("add_tool")
 			if(controlled_drone && controlled_drone.drone_status == EXODRONE_IDLE)
@@ -124,7 +123,7 @@
 				controlled_drone.remove_tool(params["tool_type"])
 			return TRUE
 		if("start_travel")
-			if(controlled_drone && !controlled_drone.travel_error())
+			if(controlled_drone && controlled_drone.can_travel())
 				var/datum/exploration_site/target_site
 				if(params["target_site"])
 					target_site = locate(params["target_site"]) in GLOB.exploration_sites
